requirements.txt
asgiref==3.7.2
Django==4.2.7 
djangorestframework==3.14.0 
grpcio==1.59.2 
grpcio-tools==1.59.2 
pip==23.3.1 
protobuf==4.25.0 
protobuf-to-dict==0.1.0 
pytz==2023.3.post1 
setuptools==68.2.2 
sqlparse==0.4.4 
tzdata==2023.3